Fixed some more signed/unsigned warnings (wow, I made a lot of these)
--HG--
branch : Trunk
extra : transplant_source : %DA%1F%5D%3E%AAl%97%B6%7Eq2%1Dp%D1%9C%C3%A0%C9v%DB

+++ TameParse/Lr/parser_state.h
@@ -103,7 +103,7 @@ namespace lr {
         static lexeme_container endOfFile((lexeme*)NULL);
         
         // Read a new symbol if necessary
-        int pos = m_LookaheadPos + offset;
+        size_t pos = m_LookaheadPos + offset;
         
         while (pos >= m_Session->m_Lookahead.size()) {
             if (!m_Session->m_EndOfFile) {
